+++ v3.0/src/components/Item/text.js
@@ -4,22 +4,19 @@ class TextList extends ItemComponent {
 	constructor(value) {
 		super(value);
 
-		const editBtn = this.item.querySelector('.item__button--edit');
-		const deleteBtn = this.item.querySelector('.item__button--delete');
-
-		editBtn.addEventListener('click', () => {
-			this.editItem(`
-			<div>
-				<input type="text" class="item__input--edit" />
-			</div>
-			<div>
-				<button type="button" class="item__button--submit">수정완료</button>
-			</div>
-			`);
-		});
-
-		deleteBtn.addEventListener('click', () => {
-			this.deleteItem();
+		this.item.addEventListener('click', (e) => {
+			if (e.target.className === 'item__button--edit') {
+				this.editItem(`
+				<div>
+					<input type="text" class="item__input--edit" />
+				</div>
+				<div>
+					<button type="button" class="item__button--submit">수정완료</button>
+				</div>
+				`);
+			} else if (e.target.className === 'item__button--delete') {
+				this.deleteItem();
+			}
 		});
 	}
